Restore old public API
I'll revert this in the next chunk of work.
